refactor: enforce hexagonal boundaries
