+++ src/Adapter/WordpressCron.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace N3XT0R\XPub\Adapter;
 
 use DI\Container;
-use N3XT0R\XPub\Application\Factory\PublisherFactory;
+use N3XT0R\XPub\Infrastructure\Publishers\PublisherFactory;
 use N3XT0R\XPub\Application\Service\Queue\JobRunner;
 use N3XT0R\XPub\Domain\Hook\FilterDispatcherInterface;
 use N3XT0R\XPub\Infrastructure\DI\ContainerProvider;
