+++ src/Application/Service/Queue/JobRunner.php
@@ -8,6 +8,7 @@ use DateTimeImmutable;
 use N3XT0R\XPub\Application\Publisher\PublisherSelector;
 use N3XT0R\XPub\Domain\Contracts\Factory\ArticleFactoryInterface;
 use N3XT0R\XPub\Domain\Contracts\QueueRepositoryInterface;
+use N3XT0R\XPub\Domain\Repository\PostStatusRepositoryInterface;
 use Psr\Log\LoggerInterface;
 
 class JobRunner
@@ -16,6 +17,7 @@ class JobRunner
         private readonly QueueRepositoryInterface $queue,
         private readonly PublisherSelector $publisherSelector,
         private readonly ArticleFactoryInterface $articleFactory,
+        private readonly PostStatusRepositoryInterface $postStatusRepository,
         private readonly ?LoggerInterface $logger = null
     ) {
     }
@@ -30,7 +32,7 @@ class JobRunner
                 $article = $this->articleFactory->fromArray($job->payload);
 
                 // Skip if article is not published
-                if (!$this->isPublishedAndNotOutdated($article->postId)) {
+                if (!$this->postStatusRepository->isPublishedAndNotOutdated($article->postId)) {
                     $this->logger?->info("Skipping job {$job->id} for unpublished post {$article->postId}");
                     continue;
                 }
@@ -49,38 +51,5 @@ class JobRunner
         }
     }
 
-    /**
-     * Checks whether a given WordPress post is published and has no newer unpublished revision.
-     *
-     * This ensures that only the latest published version is processed.
-     * If a newer revision exists (e.g. a draft or scheduled update), the method returns false.
-     *
-     * @param  int  $postId  The ID of the parent (published) post.
-     * @return bool True if the post is published and not outdated by a newer unpublished revision; false otherwise.
-     */
-    private function isPublishedAndNotOutdated(int $postId): bool
-    {
-        $post = get_post($postId);
-
-        if (!$post instanceof \WP_Post || $post->post_status !== 'publish') {
-            $this->logger?->debug("Post {$postId} is not published");
-            return false;
-        }
-
-        $revisions = wp_get_post_revisions($postId);
-
-        foreach ($revisions as $rev) {
-            // Check if revision is newer and not published (i.e., still being edited)
-            if (
-                $rev->post_modified_gmt > $post->post_modified_gmt &&
-                $rev->post_status !== 'publish'
-            ) {
-                $this->logger?->debug("Post {$postId} has newer unpublished revision {$rev->ID}");
-                return false;
-            }
-        }
-
-        return true;
-    }
 
 }
