+++ src/Domain/Hook/HookDispatcherInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace N3XT0R\XPub\Domain\Hook;
 
-use N3XT0R\XPub\Infrastructure\Wordpress\Hook\HookDefinition;
+use N3XT0R\XPub\Domain\Hook\HookDefinition;
 
 interface HookDispatcherInterface
 {
