+++ src/Infrastructure/Wordpress/Hook/WordpressHookDispatcher.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace N3XT0R\XPub\Infrastructure\Wordpress\Hook;
 
 use N3XT0R\XPub\Domain\Hook\HookDispatcherInterface;
+use N3XT0R\XPub\Domain\Hook\HookDefinition;
 
 final class WordpressHookDispatcher implements HookDispatcherInterface
 {
