+++ tests/Application/Factory/PublisherFactoryTest.php
@@ -2,7 +2,7 @@
 
 namespace N3XT0R\XPub\Tests\Application\Factory;
 
-use N3XT0R\XPub\Application\Factory\PublisherFactory;
+use N3XT0R\XPub\Infrastructure\Publishers\PublisherFactory;
 use N3XT0R\XPub\Domain\Hook\FilterDispatcherInterface;
 use N3XT0R\XPub\Infrastructure\Publishers\DevToPublisher;
 use PHPUnit\Framework\TestCase;
