+++ tests/Infrastructure/Wordpress/Factory/WordpressPublisherFactoryTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace N3XT0R\XPub\Tests\Infrastructure\Wordpress\Factory;
 
-use N3XT0R\XPub\Application\Factory\PublisherFactory;
+use N3XT0R\XPub\Infrastructure\Publishers\PublisherFactory;
 use N3XT0R\XPub\Domain\Entity\Publisher;
 use N3XT0R\XPub\Infrastructure\Wordpress\Factory\WordpressPublisherFactory;
 use N3XT0R\XPub\Infrastructure\Wordpress\Hook\WordpressFilterDispatcher;
