+++ tests/Infrastructure/Wordpress/Hook/HookProviderTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace N3XT0R\XPub\Tests\Infrastructure\Wordpress\Hook;
 
-use N3XT0R\XPub\Infrastructure\Wordpress\Hook\HookDefinition;
+use N3XT0R\XPub\Domain\Hook\HookDefinition;
 use N3XT0R\XPub\Infrastructure\Wordpress\Hook\HookProvider;
 use N3XT0R\XPub\Infrastructure\Wordpress\Admin\SettingsSaveHandler;
 use N3XT0R\XPub\Infrastructure\Wordpress\Rest\OAuthController;
