+++ tests/Stubs/DummyDispatcher.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace N3XT0R\XPub\Tests\Stubs;
 
 use N3XT0R\XPub\Domain\Hook\HookDispatcherInterface;
-use N3XT0R\XPub\Infrastructure\Wordpress\Hook\HookDefinition;
+use N3XT0R\XPub\Domain\Hook\HookDefinition;
 
 final class DummyDispatcher implements HookDispatcherInterface
 {
